move test and fix
